Fixed file access type determination
In hooks.c, MACL would report that a file was both read and executed
when the file was opened for execution. This does not coincide with the
way Unix permissions work, where a file is opened for reading, writing
or executing exclusively. Now, hooks.c correctly indicates that files
opened for executing are opened only for executing, not reading.